requirements.txt
# Azure AI Foundry Workshop - Requirements
# Dependências necessárias para executar todos os laboratórios

# Core Azure AI Libraries
openai>=1.12.0
azure-ai-inference>=1.0.0b1
azure-cognitiveservices-speech>=1.34.0
azure-ai-textanalytics>=5.3.0
azure-ai-vision-imageanalysis>=1.0.0b1
azure-ai-formrecognizer>=3.3.0
azure-ai-contentsafety>=1.0.0
azure-search-documents>=11.4.0
azure-core>=1.29.0

# Semantic Kernel and AutoGen
semantic-kernel>=1.0.0
autogen-agentchat>=0.6.0
autogen-ext[openai,azure]>=0.6.0

# Environment and Configuration
python-dotenv>=1.0.0

# Data Processing and Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
statistics>=1.0.0

# Image Processing
base64>=1.0.0
mimetypes>=1.0.0

# Async Support
asyncio>=3.4.3

# Additional Utilities
typing-extensions>=4.5.0
pydantic>=2.0.0

# Optional - For Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.21.0

# Optional - For enhanced visualization
matplotlib>=3.7.0
plotly>=5.14.0
